DolphinQt: Move gamecube controller ui logic to its own widget
