+++ Source/Core/DolphinQt/Config/ControllersWindow.h
@@ -6,19 +6,9 @@
 
 #include <QDialog>
 
-#include <array>
-
 class CommonControllersWidget;
-class MappingWindow;
+class GamecubeControllersWidget;
 class QDialogButtonBox;
-class QCheckBox;
-class QComboBox;
-class QHBoxLayout;
-class QGridLayout;
-class QGroupBox;
-class QLabel;
-class QPushButton;
-class QRadioButton;
 class WiimoteControllersWidget;
 
 class ControllersWindow final : public QDialog
@@ -28,28 +18,11 @@ public:
   explicit ControllersWindow(QWidget* parent);
 
 private:
-  void OnGCTypeChanged(int state);
-  void SaveSettings();
-  void OnGCPadConfigure();
-
-  void CreateGamecubeLayout();
   void CreateMainLayout();
   void ConnectWidgets();
-  void LoadSettings();
 
-  // Main
   QDialogButtonBox* m_button_box;
-
-  // Gamecube
-  QGroupBox* m_gc_box;
-  QGridLayout* m_gc_layout;
-  std::array<QComboBox*, 4> m_gc_controller_boxes;
-  std::array<QPushButton*, 4> m_gc_buttons;
-  std::array<QHBoxLayout*, 4> m_gc_groups;
-
-  // Wii Remote
+  GamecubeControllersWidget* m_gamecube_controllers;
   WiimoteControllersWidget* m_wiimote_controllers;
-
-  // Common
   CommonControllersWidget* m_common;
 };
